# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/{network}/close/batch-plan": {
      "post": {
        "operationId": "CloseController_batchPlan",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPlanRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "One plan per address, in the same order as the request. An address that could not be read or planned safely reports as its own `blocked` plan rather than failing the call.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid network, addresses, destination, or JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network",
                                "invalid_addresses",
                                "too_many_addresses",
                                "invalid_destination",
                                "invalid_body"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unauthorized"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 100 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "payload_too_large"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The request body is not UTF-8 JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unsupported_media_type"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "rate_limited"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Plan a bounded list of accounts in one call - what each holds, its wind-down steps, and what lands at the shared destination. Read-only; no decisions are applied per address.",
        "tags": [
          "close"
        ]
      }
    },
...
  },
  "info": {
    "title": "LumenWipe API",
    "description": "Programmatic close-out of Stellar accounts. Request bodies must be JSON objects of at most 100 KB: a larger body is rejected with 413 payload_too_large, a non-JSON content type with 415 unsupported_media_type, and a body that is not a JSON object with 400 invalid_body.",
    "version": "0.1.0",
    "contact": {
      "name": "LumenWipe",
      "url": "https://lumenwipe.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "https://api.lumenwipe.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Integrator API key"
      },
      "integrator-session": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Short-lived wallet session token from POST /integrator/auth/session"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DecisionPointDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable id, e.g. \"asset:USDC-GISSUER...\"."
          },
          "type": {
            "type": "string",
            "enum": [
              "asset_disposition",
              "confirmation",
              "choice",
              "claimable_balance"
            ]
          },
          "subject": {
            "type": "object",
            "description": "The subject of this decision - shape depends on `type` (e.g. an asset code/issuer/balance for asset_disposition, a claimable balance id for claimable_balance). Freeform by design: there is one decision-point contract per `type`, not one global shape.",
            "additionalProperties": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionOptionDto"
            }
          },
          "default": {
            "type": "string",
            "description": "The option id applied if the caller never answers this decision."
          },
          "required": {
            "type": "boolean",
            "description": "Whether an answer is mandatory before the plan can proceed."
          }
        },
        "required": [
          "id",
          "type",
          "subject",
          "options",
          "default",
          "required"
        ]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "planHash": {
            "type": "string",
            "description": "Deterministic hash of this plan's content - unchanged inputs yield the same hash."
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "needs_decisions",
              "blocked",
              "complete"
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlannedStepDto"
            }
          },
          "decisionPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionPointDto"
            }
          },
          "blockers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanResponseBlockerDto"
            }
          },
          "estimate": {
            "$ref": "#/components/schemas/PlanEstimateDto"
          },
          "execution": {
            "$ref": "#/components/schemas/PlanExecutionDto"
          }
        },
        "required": [
          "planHash",
          "status",
          "steps",
          "decisionPoints",
          "blockers",
          "estimate",
          "execution"
        ]
      },
      "BatchPlanRequestDto": {
        "type": "object",
        "properties": {
          "addresses": {
            "description": "Accounts to plan (G...), up to 20 per call.",
            "example": [
              "GABC...XYZ",
              "GDEF...UVW"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "destination": {
            "type": "string",
            "description": "Shared destination for the recovered XLM (G...). Omit to preview without one.",
            "example": "GDEF...UVW"
          }
        },
        "required": [
          "addresses"
        ]
      },
      "BatchPlanResultDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "The account this result is for."
          },
          "plan": {
            "description": "This address's plan, in the same shape close/plan returns. An address that could not be read or planned safely comes back with status \"blocked\" and the failure as its own blocker, rather than failing the whole batch.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanResponseDto"
              }
            ]
          }
        },
        "required": [
          "address",
          "plan"
        ]
      },
      "BatchPlanResponseDto": {
        "type": "object",
        "properties": {
          "results": {
            "description": "One result per requested address, in the same order as the request.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchPlanResultDto"
            }
          }
        },
        "required": [
          "results"
        ]
      },
...
    }
  }
}